changed method displayRunningTime to take parameter as double instead of int

+++ PIC_Simulator/src/pic_simulator/interfaces/ModelPresenter.java
@@ -23,7 +23,7 @@ public interface ModelPresenter {
 
     public void displayAutomaticSteppingIntervall(int ms);
 
-    public void displayRunningTime(int microSeconds);
+    public void displayRunningTime(double microSeconds);
 
     public void displayOscillatorFrequency(double megaHz);
